+++ test/integration/flush_agent_spec.rb
@@ -10,7 +10,6 @@ describe 'FlushAgent' do
 
     # create agent
     result = @flush_agent.create_update('Some Flush Agent Title', 'Some flush agent description', 'http://somehost:8080')
-    expect(result.is_success?).to be(true)
     expect(result.message).to eq('Flush agent some-flush-agent created on author')
   end
 
@@ -21,13 +20,11 @@ describe 'FlushAgent' do
 
     it 'should succeed existence check' do
       result = @flush_agent.exists()
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Flush agent some-flush-agent exists on author')
     end
 
     it 'should succeed update' do
       result = @flush_agent.create_update('Some Updated Flush Agent Title', 'Some updated flush agent description', 'https://someotherhost:8081')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Flush agent some-flush-agent updated on author')
     end
 
@@ -37,11 +34,9 @@ describe 'FlushAgent' do
 
     it 'should succeed existence check' do
       result = @flush_agent.delete()
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Flush agent some-flush-agent deleted on author')
 
       result = @flush_agent.exists()
-      expect(result.is_failure?).to be(true)
       expect(result.message).to eq('Flush agent some-flush-agent not found on author')
     end