+++ test/integration/group_spec.rb
@@ -7,11 +7,9 @@ describe 'Group' do
     # ensure group doesn't exist prior to testing
     @group = @aem.group('/home/groups/s/', 'somegroup')
     result = @group.delete()
-    expect(result.is_success?).to be(true)
 
     # create group
     result = @group.create()
-    expect(result.is_success?).to be(true)
     expect(result.message).to match(/^Group somegroup created at \/home\/groups\/s\/.+/)
   end
 
@@ -22,13 +20,11 @@ describe 'Group' do
 
     it 'should succeed existence check' do
       result = @group.exists()
-      expect(result.is_success?).to be(true)
       expect(result.message).to match(/^Group somegroup exists at \/home\/groups\/s\/.+/)
     end
 
     it 'should succeed permission setting' do
       result = @group.set_permission('/etc/replication', 'read:true,modify:true')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Permission read:true,modify:true on path /etc/replication set for group somegroup')
     end
 
@@ -37,16 +33,13 @@ describe 'Group' do
       # ensure member group doesn't exist prior to testing
       member_group = @aem.group('/home/groups/s/', 'somemembergroup')
       result = member_group.delete()
-      expect(result.is_success?).to be(true)
 
       # create member group
       result = member_group.create()
-      expect(result.is_success?).to be(true)
       expect(result.message).to match(/^Group somemembergroup created at \/home\/groups\/s\/.+/)
 
       # add user as member to the group
       result = @group.add_member('somemembergroup')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('User/group somemembergroup added to group somegroup')
     end