+++ test/integration/node_spec.rb
@@ -8,7 +8,6 @@ describe 'Node' do
     @node = @aem.node('/apps/system/', 'somefolder')
     @node.delete()
     result = @node.exists()
-    expect(result.is_failure?).to be(true)
     expect(result.message).to eq('Node apps/system/somefolder not found')
   end
 
@@ -19,14 +18,12 @@ describe 'Node' do
 
     it 'should succeed when node does not exist' do
       result = @node.create('sling:Folder')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder created')
     end
 
     it 'should raise error when node already exists' do
 
       result = @node.create('sling:Folder')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder created')
 
       # create the same node the second time
@@ -40,17 +37,14 @@ describe 'Node' do
     it 'should succeed existence check when node already exists' do
       # node does not exist
       result = @node.exists()
-      expect(result.is_failure?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder not found')
 
       # create node
       result = @node.create('sling:Folder')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder created')
 
       # node should exist
       result = @node.exists()
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder exists')
     end
 
@@ -62,11 +56,9 @@ describe 'Node' do
 
       # ensure node exists prior to deletion
       result = @node.create('sling:Folder')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder created')
 
       result = @node.delete()
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder deleted')
     end
 
@@ -78,11 +70,9 @@ describe 'Node' do
 
       # ensure node exists prior to deletion
       result = @node.create('sling:Folder')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder created')
 
       result = @node.exists()
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Node apps/system/somefolder exists')
     end