+++ test/integration/path_spec.rb
@@ -13,7 +13,6 @@ describe 'Path' do
     it 'should succeed when path exists' do
       path = @aem.path('/etc/designs/cloudservices')
       result = path.activate(true, false)
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Path /etc/designs/cloudservices activated')
     end
 
@@ -21,7 +20,6 @@ describe 'Path' do
     it 'should still succeed when path does not exist' do
       path = @aem.path('/some/inexisting/path')
       result = path.activate(true, false)
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Path /some/inexisting/path activated')
     end