+++ test/integration/replication_agent_spec.rb
@@ -10,7 +10,6 @@ describe 'ReplicationAgent' do
 
     # create agent
     result = @replication_agent.create_update('Some replication Agent Title', 'Some replication agent description', 'http://somehost:8080')
-    expect(result.is_success?).to be(true)
     expect(result.message).to eq('Replication agent some-replication-agent created on author')
   end
 
@@ -21,13 +20,11 @@ describe 'ReplicationAgent' do
 
     it 'should succeed existence check' do
       result = @replication_agent.exists()
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Replication agent some-replication-agent exists on author')
     end
 
     it 'should succeed update' do
       result = @replication_agent.create_update('Some Updated replication Agent Title', 'Some updated replication agent description', 'https://someotherhost:8081')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Replication agent some-replication-agent updated on author')
     end
 
@@ -37,11 +34,9 @@ describe 'ReplicationAgent' do
 
     it 'should succeed existence check' do
       result = @replication_agent.delete()
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Replication agent some-replication-agent deleted on author')
 
       result = @replication_agent.exists()
-      expect(result.is_failure?).to be(true)
       expect(result.message).to eq('Replication agent some-replication-agent not found on author')
     end