+++ test/integration/user_spec.rb
@@ -7,11 +7,9 @@ describe 'User' do
     # ensure user doesn't exist prior to testing
     @user = @aem.user('/home/users/s/', 'someuser')
     result = @user.delete()
-    expect(result.is_success?).to be(true)
 
     # create user
     result = @user.create('somepassword')
-    expect(result.is_success?).to be(true)
     expect(result.message).to match(/^User someuser created at \/home\/users\/s\/.+/)
   end
 
@@ -22,19 +20,16 @@ describe 'User' do
 
     it 'should succeed existence check' do
       result = @user.exists()
-      expect(result.is_success?).to be(true)
       expect(result.message).to match(/^User someuser exists at \/home\/users\/s\/.+/)
     end
 
     it 'should succeed permission setting' do
       result = @user.set_permission('/etc/replication', 'read:true,modify:true')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Permission read:true,modify:true on path /etc/replication set for user someuser')
     end
 
     it 'should succeed admin password change' do
       result = @user.change_password('admin', 'admin')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Logged in user password changed')
     end
 
@@ -49,7 +44,6 @@ describe 'User' do
       })
       user = aem.user('/home/users/s/', 'someuser')
       result = user.change_password('somepassword', 'somenewpassword')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('Logged in user password changed')
 
     end
@@ -59,16 +53,13 @@ describe 'User' do
       # ensure group doesn't exist prior to testing
       group = @aem.group('/home/groups/s/', 'somegroup')
       result = group.delete()
-      expect(result.is_success?).to be(true)
 
       # create group
       result = group.create()
-      expect(result.is_success?).to be(true)
       expect(result.message).to match(/^Group somegroup created at \/home\/groups\/s\/.+/)
 
       # create group
       result = @user.add_to_group('/home/groups/s/', 'somegroup')
-      expect(result.is_success?).to be(true)
       expect(result.message).to eq('User/group someuser added to group somegroup')
 
     end