+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 1.0.0
 * Fix replication agent creation failure #1
 * Unsuccessful action should now raise RubyAem::Error, replacing failure result
+* Remove status attribute from RubyAem::Result
+* Add response attribute to RubyAem::Result
 * Move resource classes to RubyAem::Resources module
 
 ### 0.9.0